Move saving/loading functions and files into data and make it a hidden module
