--- docs/api/data.rst
@@ -1,5 +1,0 @@
-Data
-====
-
-.. automodule:: huracanpy.data.ibtracs
-    :members:
