--- huracanpy/data/__init__.py
@@ -1,5 +1,0 @@
-"""huracanpy module for shared data"""
-
-__all__ = ["ibtracs"]
-
-from . import ibtracs
