+++ pyproject.toml
@@ -41,7 +41,7 @@ requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.package-data]
-huracanpy = ["example_data/*", "data/_ibtracs_files/*"]
+huracanpy = ["_data/example_data/*", "_data/_ibtracs_files/*"]
 
 [tool.ruff]
 extend-include = ["*.ipynb"]
